Fixed bug with file inclusion, updated Cloudflare help texts, added encryption to options
